Prometheus public IP exporter
This exporter provides the current public IP address of the system or
the network respectively the exporter is running on. The IP address is
stored in a label while the metric itself gives the age of the IP
address counted from the start of the exporter or the last renewal
timestamp.

Issue: #main