Sub-system configuration option
To allow calculation of certain values like e.g., Usage (P_Usage, or
Total Production (P_PV) it is now possible to add additional inverters
as subsystems into the configuration per system:

- Each System (the total of all inverters connected to each other) has
  one controlling inverter (Controller) connected to the smart meter
  and exposing all metered values.

- Each controller may have a number of additional inverters
  (Subsystems) attached to it.

- Each Subsystem is queried for their production value (P_PV) directly
  from the query to the Controller.

- All P_PV values from the Subsystems are added to the P_PV value
  derived from the Controller making this value reflecting the total
  production of the System.

- Additionally all P_PV values derived from the Subsystems are added
  as additional metrics (P_PV_<subsystem_num>) to the output of the
  Controller.

- Optionally each Subsystem may be queried as an individual Prometheus
  target with no Subsystems attached.

With this setup the exporter is able to calculate several metrics
directly. Such metrics are the overall Consumption of the site
(P_Usage) as well as to provide seperate metrics for power to and from
Grid based on the sign of the value (- for to Grid and + from Grid).

It is easier to do such calculations in the exporter directly compared
to calculating these numbers in Grafana.

Issue: #main